Fix chimney sealing, stone charging and the thermometer layout
Three defects found in play-testing:

- A chimney led outside made the room read as leaking. The flue was
  seeded as part of the stove, so the flood fill started in the open air
  above the roof and overflowed the volume cap. Only the firebox and its
  masonry seed the fill now; the flue walls the room in as an ordinary
  solid block.
- Stones charged one at a time. Heat was an integer, so a three-a-second
  fire could not be split across eight stones and all of it landed in
  one. Heat is a float now and both charge and release share evenly.
- The thermometer's rows ran off the panel. Label and value now share a
  line above a full-width bar, and the panel grew to 176x132.

+++ src/main/java/com/banya/item/BanyaStoneItem.java
@@ -31,8 +31,8 @@ public class BanyaStoneItem extends Item {
         if (level.isClientSide() || level.getGameTime() % COOL_INTERVAL_TICKS != 0) {
             return;
         }
-        int heat = StoveStones.heatOf(stack);
-        if (heat > 0) {
+        float heat = StoveStones.heatOf(stack);
+        if (heat > 0.0F) {
             // Out of the stove it gives its warmth up to the room, slowly enough to be carried.
             StoveStones.setHeat(stack, heat - Config.STONE_COOLING_PER_STEP.get());
         }
@@ -40,15 +40,15 @@ public class BanyaStoneItem extends Item {
 
     @Override
     public boolean isBarVisible(ItemStack stack) {
-        return StoveStones.heatOf(stack) > 0;
+        return StoveStones.heatOf(stack) > 0.0F;
     }
 
     @Override
     public int getBarWidth(ItemStack stack) {
-        int capacity = Math.max(1, StoveStones.capacityOf(stack));
+        float capacity = Math.max(1, StoveStones.capacityOf(stack));
         // A big stove banks more into a stone than the stone holds on its own, so a freshly pulled
         // one can read over full. Clamp, or the bar runs off the end of the icon.
-        float filled = Math.min(1.0F, StoveStones.heatOf(stack) / (float) capacity);
+        float filled = Math.min(1.0F, StoveStones.heatOf(stack) / capacity);
         return Math.clamp(Math.round(BAR_SEGMENTS * filled), 1, BAR_SEGMENTS);
     }
